list_models.py
from google import genai
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")

client = genai.Client(api_key=api_key)

print("可用模型列表:")
for model in client.models.list():
    print(f"- {model.name} ({model.display_name})")
