requirements.txt
playwright
python-dotenv
requests
loguru
google-genai
Pillow
flask
flask-cors
